+++ seantis/dir/events/setuphandler.py
@@ -1,0 +1,9 @@
+from seantis.plonetools import setuphandlers
+
+indexes = [
+    ('source', 'FieldIndex'),
+]
+
+
+def import_indexes(context):
+    setuphandlers.import_indexes('seantis.dir.events', indexes, context)