+++ docs/HISTORY.rst
@@ -5,7 +5,8 @@ Changelog
 1.3.5 (unreleased)
 ~~~~~~~~~~~~~~~~~~
 
-Nothing yet.
+- Set up indexes on installation. Fixes #84.
+  [msom]
 
 
 1.3.4 (2014-11-27)